requirements.txt
pytesseract
pillow
mysql-connector-python
gunicorn
flask
opencv-python
werkzeug